Add "measure" command to measure execution time
To measure the execution time of a command more precisely, this commit
adds a new command, as indicated in the title, allowing developers to
use it and to measure a command's execution time.

The new command can execute a given command multiple times if the first
argument is greater than 1, and then output the timing for each
execution. Therefore, developers can use these outputs for analysis and
and statistics.

Change-Id: I1ec297aa7c785327810e09010d0a4ccc5087a43e